requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
gunicorn==23.0.0
Pillow
pillow-heif